Message ignoring logic and regex fail number ???

+++ bot/nepeatbot/plugins/plugin_moderation.py
@@ -38,11 +38,17 @@ class ModerationPlugin(PluginBase):
 
             for messages in zip_longest(*(iter(removed),) * 100):
                 messages = [message for message in messages if message]
+                await self.redis.sadd("ignored:{}".format(message.server.id), [m.id for m in messages])
                 await self.bot.delete_messages(messages)
+
+            await self.redis.expire("ignored:{}".format(message.server.id), 120)
 
         return Message("{} messages removed!".format(deleted))
 
-    @command("purgechan(?:\s(\d+))?")
+    @command(patterns=[
+        "purgechan (\d+)",
+        "purgechan"
+    ])
     async def cmd_purge_chan(self, message, args):
         try:
             limit = int(args[0])
@@ -59,10 +65,16 @@ class ModerationPlugin(PluginBase):
         for messages in zip_longest(*(iter(removed),) * 100):
             messages = [message for message in messages if message]
             await self.bot.delete_messages(messages)
+            await self.redis.sadd("ignored:{}".format(message.server.id), [m.id for m in messages])
+
+        await self.redis.expire("ignored:{}".format(message.server.id), 120)
 
         return Message("{} messages removed!".format(len(removed)))
 
     @command("remove (\d+)")
     async def cmd_remove(self, message, args):
         message = await self.bot.get_message(message.channel, args[0])
-        await self.bot.delete_message(message)
+        if message:
+            await self.redis.sadd("ignored:{}".format(message.server.id), [message.id])
+            await self.bot.delete_message(message)
+
